fix more 64bit warnings and fix some minor issues caused by prev fixes
..and test for script init failure
..this makes VP fully functional (except for the scripting part, and as far as my testing goes) as 64bit compile

git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@1164 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ pin/player.cpp
@@ -3189,7 +3189,7 @@ INT_PTR CALLBACK DebuggerProc(HWND hwndDlg, UINT uMsg, WPARAM wParam, LPARAM lPa
                         {
                            HWND hwndControl = GetDlgItem(hwndDlg, IDC_BALL_THROWING);
                            size_t checked = SendMessage(hwndControl, BM_GETCHECK, 0, 0);
-                           g_pplayer->m_fThrowBalls = checked;
+                           g_pplayer->m_fThrowBalls = (int)checked;
                            break;
                         }
 							}
